pyproject.toml
[project]
name = "skills-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "pdfplumber>=0.11.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "tqdm>=4.67.1",
]
